+++ uart_Init.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif
 
 extern void UART_Init(uint32_t baudRate);
-extern void UART_Write_char(uint16_t iData);
+extern void UART_Write_char(uint8_t iData);
 extern void UART_Write_String(uint8_t cText[]);
 extern char UART_Read(void);
 extern bool UART_Data_available(void);
